requirements.txt
-e ./texasholdem    
openai>=1.0.0
python-dotenv>=1.0.0 
maturin>=0.16.0
transformers>=4.45.2
torch==2.2.2 
torchvision==0.17.2
torchaudio==2.2.2
accelerate